Beginning to incorporate transaction logic
So much of the old placement was tied to sqlalchemy: not just the DB
interaction code, but the whole system of the context objects carrying
around sessions and transactions. This is the first step in making that
work with Neo4j.

I've switched from using py2neo as the Python wrapper to the
officially-supported python-neo4j client. Besides the official support,
it has much cleaner transaction handling. These changes are going to
require re-writing all of the DB interaction code to a) ensure that
context objects are available wherever needed, and b) the return values
from the new calls are handled correctly. py2neo returned Node objects,
which then had to be 'pythonized' to work, while neo4j return straight
Python dicts.