Transactions are *almost* working
Reworked the db transaction decorator for py2neo. It's working for the
most part, but there are still issues when nesting several transactions
failing to roll them all back when an exception is raised. So I'm
committing this now, and may try again with the python-neo4j wrapper
instead of py2neo to see if it behaves better.