Updated the merged code from upstream
Pulling the changes from openstack/placement into replacement meant that
a lot of the old model crept back in. There was also a large refactoring
where much of the resource_provider object code was moved into the new
research_context object, so that had to be reconciled.

Still haven't gotten transactions working, though.